Move Django SolrClient to a separate file
