+++ parasol/solr/__init__.py
@@ -1,9 +1,3 @@
-try:
-    import django
-except ImportError:
-    django = None
 
 from parasol.solr.client import SolrClient
 
-if django:
-    from parasol.solr.client import DjangoSolrClient
